math libm: (u16)angle << 2 in fn_80293EAC for clrlslwi peephole
s16 reduced = (u16)angle << 2 makes MWCC use clrlslwi instead of slwi,
matching target. fn_80293EAC 83.92% -> 84.76%.

+++ src/main/unknown/autos/placeholder_80293464.c
@@ -132,7 +132,7 @@ float fn_80293D0C(int angle) {
 }
 
 float fn_80293EAC(int angle) {
-    s16 reduced = angle << 2;
+    s16 reduced = (u16)angle << 2;
     float y = fn_80291E08(&reduced);
     float y2 = y * y;
